requirements.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
braceexpand==0.1.7
brotli==1.2.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
Cython==3.2.4
decord==0.6.0
distro==1.9.0
easydict==1.13
einops==0.8.1
fairscale==0.4.13
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
fire==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
hub-sdk==0.0.24
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
insightface==0.7.3
iopath==0.1.10
jax==0.9.0
jaxlib==0.9.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lapx==0.9.4
lazy_loader==0.4
lightning-utilities==0.15.2
lmdeploy==0.11.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.13
ml_dtypes==0.5.4
mmengine-lite==0.10.7
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
narwhals==2.15.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openai==2.15.0
openai-harmony==0.0.8
opencv-contrib-python==4.13.0.90
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.11.5
packaging==26.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
peft==0.14.0
pillow==12.1.0
platformdirs==4.5.1
portalocker==3.2.0
prettytable==3.17.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.1
py-cpuinfo==9.0.0
pyarrow==23.0.0
pybase64==1.4.3
pycocoevalcap==1.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytorchcv==0.0.74
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.4
starlette==0.50.0
streamlit==1.53.1
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2026.1.14
tiktoken==0.12.0
timm==0.6.13
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.6
triton==3.4.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.1.0
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.3.0
webdataset==1.0.2
wheel==0.45.1
xgrammar==0.1.31
yapf==0.43.0
yarl==1.22.0
yt-dlp==2025.12.8
